internal core include restructuring
